Fixed the GUI duplication bug, and added a status label!
I'm assuming because I edited the name of the 'GUI' folder into 'gui'. However, something weird happened and had duplicated both of the files. Causing some conflict with the class_name as Godot found two of the same files in different folders (even if the 'GUI' doesn't exist in the file explorer. This is found in the '1d27958' commit.

Other than that, I added the implemented the status text for when the player take a screenshot, and set to fullscreen or windowed.