Fix cheese not spawning anymore when eaten
This is because whenever a cheese is eaten, it doesn't update the amount of cheese existed on screen. The cheese manager thought that there is still 46 cheese even if there is no cheese on screen.

By connecting each cheese's tree_exited signal. We can use that to update the cheese_amount whenever possible.