Trying to add fullscreen. Error in description.
It's constantly popping up after "implementing the fullscreen shortcut".

The cheese_manager's get_cheese_amount_present() is causing problems. It's trying to get a tree that doesn't exists. This might because tree_exited() also runs whenever the game closes. But since the tree doesn't exist anymore, it starts to cry.

Might create a safer signal that handles this update.