Massive update.
This contains the latest work that I've done on the game. It's a bit messy as I end up doing a lot of hacks to make the game more optimize. I was able to get it up to 5000 cheeses by reducing the audio objects being created in closed areas.

I could have done more to make it organized as I didn't expect to share this code all around.

This update doesn't reflect what's shown in itch.io as of Sept. 14, 2023.